requirements.txt
fastapi
uvicorn
pydantic
requests
streamlit
python-dotenv
pytest
pytest-mock
httpx
langchain
langchain-groq
groq